Move test classes into separate files
